shell,nix: set original PATH at start of init file (#57)
This fixes a bug that causes the devbox shell's init to fail when the
user's rcfile tries to call a command outside of devbox. For example,
the following line in a ~/.bashrc will trigger an error if the devbox
doesn't have Go installed:

	export PATH="$(go env GOPATH):$PATH"

This is because we're adding the ORIGINAL_PATH at the end of their shell
init script, which means that anything before that only see Nix
packages.

Fix this by adding a pre-init hook that lets us restore the PATH before
the user's rcfile, while still forcing Nix packages to come first in
the post-init hook.

---

In the process of testing and debugging this PR, I added a whole bunch
of debug logging. It should probably be a separate commit, but I'm
hesitant to touch anything now that things are working.