boxcli: add basic tests for devbox shell (#83)
Add tests that check that `devbox shell` launches successfully on the
platforms (Linux and macOS) and shells (bash, dash, and zsh) that we
support. Specifically, each test runs the equivalent of typing the
following into your terminal:

	cd `mktemp -d`
	devbox init
	devbox add hello
	devbox shell
	hello
	exit
	exit

The bulk of this change is the addition of some test helpers that handle
launching shells and interacting with them. For now they live in the
same file as the `boxcli` shell tests, but we can move them into their
own `shelltest` package later on if necessary. With these test helpers
in place, it should be a lot easier to add tests for more complicated
scenarios.

In order to run these tests in CI, the GitHub test jobs now install Nix
along with some additional shells. They also run on both ubuntu-latest
and macos-12 to make sure things work on both systems. The lint job also
runs on Ubuntu and macOS to make sure that we don't inadvertently add
platform-specific code that won't compile.