Add Analyzer framework (#99)
## Summary
Adds an `Analyzer` to the `plansdk`. Analyzers can be initialized on a
root directory, and provide lots of utility functions to then understand
(or analyze) the source code present in that root directory. Analyzer
implements several functions that are really handy when implementing all
sorts of planners.

This PR just introduces the library; future PRs would start using the
library in the implementation of planners.

## How was it tested?
`go test -v ./...`