nix: allow user's shellrc to update PATH; add tests for shellrc template (#106)
At the end of the devbox shellrc file, we're setting:

	export PATH="$PURE_NIX_PATH:$ORIGINAL_PATH"

This will overwrite anything that the user's shellrc might've added to
PATH. Change ORIGINAL_PATH to PATH to preserve those changes. There's
also no need for ORIGINAL_PATH to be an environment variable, so set it
directly in the template instead.

To prevent regressions, some tests that generate a devbox shellrc file
and compare it to an expected golden file. This ensures we don't
unknowingly change the way the shellrc is generated.

If we do intentionally change the devbox shellrc, the tests can
automatically update the golden files with -update:

	go test ./nix -update

It's up to the author of the change to verify that any new shellrc files
still work in a real shell.