[CSharp Planner] part 1: devbox shell (#120)
## Summary

This PR adds a C# Planner.

Do note that the nix package we use is that of the Dot Net Framework.
And a
particular Dot Net Framework version (say, 6) can support its latest C#
version
and older versions. For example, dotnet-6 supports C# 10 and C# 9, and
possibly older C# as well.

So, this Planner is in essence a DotNetPlanner, more than a
CSharpPlanner. When
we look to add support for other DotNet languages like FSharp, we can
either
reuse this planner by renaming to DotNetPlanner, or have some mixin
abstraction
for CSharpPlanner and FSharpPlanner to share logic.


## How was it tested?

added to `testdata/csharp`.

In each folder, I did:
```
> devbox shell

# verify version is correct
> dotnet --version

# expect Hello World
> dotnet run
```