devbox shell -- cmd (#158)
## Summary
Simple implementation of `devbox shell -- <cmd>`. I struggled
integrating this with all the other stuff that is happening with `devbox
shell`. Would like thoughts on whether this is a good enough for an MVP
of the feature, and later we improve it, or whether we think integrating
with the other `devbox shell` logic is a must have.

## How was it tested?
Ran a few commands with `devbox shell --`