devbox: add PackageStore and Package for querying /nix/store (#270)
This is the beginnings of a new API that allows Devbox to query the Nix
store without needing to call out to the nix command. So far, it can
list installed packages and determine their dependencies.

`go doc` summary of the new types:

	type PackageStore struct {
		fs.FS
		// Has unexported fields.
	}
	    PackageStore is a file system containing packages that are compatible
	    with Devbox.

	func LocalNixStore(path string) *PackageStore
	func (p *PackageStore) Package(name string) (Package, error)

	type Package struct {
		fs.FS

		// Name is the full name of the package that is unique to its store.
		Name string

		// DirectDependencies are the other packages in the store that this
		// package depends on. It does not contain transitive dependencies.
		DirectDependencies []Package
	}
	    Package is a file system that contains a package's files and metadata.

The go.mod + go.sum changes are from a `go get -u ./... && go mod tidy`.