[midcobra] add executionID (#358)
## Summary

**Motivation**
The sshshim command path bypasses the existing cobra stack. We need to
add some basic telemetry around error tracking so we have visibility
into errors users may run into (due to our programming mistakes).

We would like to reuse the existing `midcobra` abstraction which has a
`telemetry` middleware that does both sentry and segment tracking.
However, for sshshim we desire just sentry, not segment. So we want to:
1. split telemetry middleware into sentry and segment middlewares
2. set a common sentryID for both sentry and segment to correlate the
corresponding events/data.

**Implemention**
This PR adds an executionID to `midCobra.Middleware`.

We use this executionID in two ways:

1. We set executionID in both sentry.EventID and segment's
sentry_event_id
2. For an error and stack trace reported when `DEVBOX_DEBUG=1`, we also
print this executionID.
If a user reports a reproducible error, then we can ask them to enable
DEVBOX_DEBUG to reproduce
it and report the executionID to us. We can then look up the
corresponding events in both
sentry and segment to construct an understanding of the problem the user
is facing.


In the next PRs, I will:
1. split telemetry into sentry and segment middleware
2. call the sentry middleware for the sshshim command path

## How was it tested?

Set up:
1. via code changes, forced the `midcobra.telemetry` middleware to be
enabled.
2. added a print statement where the segment.sentry_event_id is set to
print
both the executionID and the eventID returned by sentry's
CaptureException.
3. Manually inserted an error in `devbox.Open` and ran a CLI command to
trigger it.

Observed:
1. executionID and eventID from sentry's CaptureException are the same.
2. the executionID is printed prior to the stack trace of the error.