Fix devbox run for fish and possibly other shells (#464)
## Summary

For shells that aren't fully detected and for which we don't have custom
logic (like fish), we fallback to just running the shell. `devbox run`
was failing to attempt to run any command in this case. This PR makes it
so that we at least attempt to run the command. It may or may not work,
depending on the shell, but it does work with fish.

This is a bit of a hacky fix to get fish to work. Longer-term I want to
re-implement devbox run so it doesn't depend on specific shells and
rcfiles, so we can avoid these compatibility problems altogether.

## How was it tested?
```
SHELL=fish ./devbox run <script>
```