Experimental impl of nixless shell (#457)
## Summary
This is a simplistic implementation of `devbox shell` that launches a
shell directly, instead of calling `nix-shell`. Instead it uses `nix
print-dev-env` to construct the environment needed to replicate a nix
shell, and launches the shell directly. As with `devbox shell`, it lets
some parts of the user's environment "leak" into the shell, in the same
ways that the current implementation of `devbox shell` does. The
behavior is behind a feature flag, so this PR should be a no-op for any
user that doesn't explicitly enable the feature.

The current implementation relies on exporting the variables in the
shellrc file, which I'd like to change in the future (will probably just
change it on this PR itself).

## How was it tested?
Tried it with a few shells and verified that the PATH and other vars
looked reasonable, and that calls to nix packages resolved to the
packages in the nix store.
```
DEVBOX_FEATURE_NIXLESS_SHELL=1 SHELL=[zsh bash sh] ./devbox shell

DEVBOX_FEATURE_NIXLESS_SHELL=1 ./devbox shell -- hello
```