[global] MVP of devbox global/home (#599)
## Summary

This is MVP of devbox global/home. It uses the default nix profile. Since the default nix profile is already in path, we don't need anything different for that. This implements:

```bash
devbox global add <pkg>...
devbox global rm <pkg>...
devbox global list
```

A few caveats:
* This approach doesn't match devbox shell, but is simpler to implement and doesn't require making their user change their rc files. 
* This won't run setup-hooks.
* This causes anything in nix default profile to "leak" into devbox project shells. (we could fix this by using non-default profile, but I'm not sure this makes sense).
* The `global` command is hidden.

I think this approach certainly makes trade-offs, but gets most of the benefits fast and with a pretty nice UX. The setup-hooks issue is something we could fix later on.

Thoughts?

## How was it tested?

<img width="425" alt="image" src="https://user-images.githubusercontent.com/544948/217477634-b03c5587-307c-42e4-b442-c2a315cce4b4.png">