Refactor env vars as map[string]string instead of []string (#621)
## Summary
This makes it easier to manipulate vars in functions (like
`computeNixEnv`). On the flip side, it does mean we need to convert them
to `[]string` in order to pass them to a `Command`, which is a bit
annoying, but I think it's reasonable. We can introduce a helper
function `asPairs(map[string]string)` or something like that later on if
we feel the need to clean up.

## How was it tested?
CICD