[config] Set PATH as ENV instead of hook (#786)
## Summary

This change improves devbox.json and plugin jsons by allowing `PATH` to
be set as an env variable instead of a hook.

For example, it allows `"PATH": "$PWD/dist:$PATH"`. 

In general any environment variable that is defined at the time of
parsing is accessible. This means for plugins, the HOST environment is
accessible, and for `devbox.json` the HOST environment and also the
plugin environment is accessible. In practice this allows the user to
override any plugin environment using an existing variable as base. On
the other hand, the plugin environment variables will not take into
account any changes that are made in the `devbox.json`

Some benefits of this change:

* We no longer need to export init hooks. One of the main use cases of
init hooks is now moved to env instead.
* `refresh` no longer breaks changes to PATH. This is because init hooks
don't get run by refresh, but the environment is reset correctly.

Related fixes:

* Removed code that was adding virtenv/<pkg>/bin to path. This is not
used.
* Removed code that was creating `env` files for each plugin. This is no
longer used.
* Added `hash -r` to refresh functions
* No longer fetch plugin env vars when starting a service (this was only
needed previously when adding a package while in shell would add it to
profile but would not add env vars to current shell. Today `refresh`
does this)

## How was it tested?

```bash
devbox shell
echo $PATH
devbox add ruby 
refresh
echo PATH
```