[flakes] Add ability to add custom local flakes to devbox project (#866)
## Summary

This allows users to specify a local flake as a package in devbox.json.
It can be added by using the following syntax:

`path:path/to/flake#output`

Where `output` can be a package that is part of `packages` or
`legacyPackages` output. ([legacyPackages is just a hack that nixpkgs
uses for
performance](https://github.com/NixOS/nixpkgs/blob/master/flake.nix#L47))

The path to flake can be relative or absolute. If a relative path is
used, it is relative to the project dir.

TODOs: (in this PR)

- [x] devbox add doesn't work because we still look for package in
nixpkgs
- [x] devbox rm doesn't work because NixProfileListItem.AttributePath is
broken for these. (see TODO in code)
- [x] NixProfileListItem.PackageName and callsites needs to modified to
ensure we don't re-install the flake output to the profile every time
(cc: @savil for visibility)

TODOs: (in follow up)

- [ ] Allow remote (github) flakes 
- [ ] Possible improvement: allow multiple outputs? (e.g.
`path:path/to/flake#output1,output2`). This is not nix syntax so not
sure we should support. cc: @Lagoja

## How was it tested?

see flakes/php example.

`devbox run "php -m | grep memcached"`