[version] part 1: update command for devbox and launcher updates (#965)
**Motivation:**
We want to disable auto-update for devbox CLIs. This will minimize
unexpected interruptions for our users.

To do so, we are going to do the following:

1. Update `launch.sh` file in `axiom` repo to write new versions to
`{XDG_CACHE_HOME}/devbox/available-version`.
https://github.com/jetpack-io/axiom/pull/3150
2. Change `devbox version update` to update the launcher (if needed) and
else, update the CLI binary. (this PR).
3. Change the `vercheck.CheckVersion` that runs in `boxcli/root.go` on
every command to check if the launcher or CLI binary versions are
out-of-date, and to print a notice if so. (reviewed in #966, but merged
into this PR)

We will cherry-pick ~these two PRs~ this PR for the next release so
auto-update stops for users.

**Implementation:**
This PR enables `jetpack version update` to:

1. Update the launcher: We detect if the launcher version is outdated,
and update it, if needed. This also updates the devbox CLI binary as a
side-effect. This requires `sudo` so we only do it, if necessary.

2. Update just the devbox CLI binary: If the launcher is up-to-date,
then we update just the devbox CLI binary.

3. Remove dependency on new code like `env` package to minimize
cherry-picking merge conflicts.

TODO:
- [x] confirm that we can switch the `devbox version` operation after an
update to `StdOut`. I feel this is better so user is informed about the
new version.
cc @mikeland86 for the above question.

Will only land this after the launch.sh file is updated in
https://github.com/jetpack-io/axiom/pull/3150

Did some basic testing:

hardcoded `true` to ensure `selfUpdateLauncher` runs, and did `devbox
version update`:

<img width="742" alt="Screenshot 2023-05-01 at 8 57 46 PM"
src="https://user-images.githubusercontent.com/676452/235576315-03069599-a469-4f3a-90c2-cc6a58ec648a.png">

regular `devbox version update`
<img width="332" alt="Screenshot 2023-05-01 at 8 48 20 PM"
src="https://user-images.githubusercontent.com/676452/235575369-f2a70299-4a3a-4b41-80ec-ecedb730ad0b.png">