[direnv] copy bash, fish, zsh shell code from direnv (#1156)
## Summary

**Overall Motivation**
We've been having issues with the bin-wrappers, and want to explore an
alternate approach. Bin-wrappers ensure that the shell environment is
completely updated prior to invoking the binary, via running `devbox
shellenv` if needed.

This alternate approach takes inspiration from direnv. It will seek to
register a prompt `hook` in the shell. The hook will execute a shell
function that will `export` the shell environment to update it, prior to
displaying the shell prompt.

To punt on how to transition existing devbox commands, we're introducing
`devbox hook` and `devbox export` as distinct commands, analogous to
`direnv hook` and `direnv export`.

**This PR's motivation**

As part of implementing this, we'd like to make use of direnv's
battle-tested shell code where we can. Direnv has a nice shell
abstraction for handling the hook and environment-variables. We could
adopt this abstraction to better support shell code we generate beyond
this immediate use-case of prompt-hooks.

So, we:
1. Copy bash, fish, zsh `Shell` code from direnv. Thank you 🙏 
2. Add ksh, posix, unknownSh structs. I filled in the struct functions
as placeholders for now.

I call the package `shenv` since it handles the "shell environment". I
stayed away from "shellenv" to avoid confusion with our existing
"shellenv" command's operation.



## How was it tested?

Have _not_ tested. Subsequent PR will test and correct as needed.