[lockfile] When updating store-paths, we should update all fields (#1530)
## Summary



When we update the store-paths in the lockfile, we should also update
the other fields, including "resolved" for the nixpkgs commit hash, and
the modified-time.

This way the exact same derivation is used whether installing the binary
directly via store-path, or via relying on the derivation in the
package's "resolved" field. This ensures parity for the following
scenarios:
1. A team has a devbox project. Some team members are using nix >= 2.17,
while others are using older nix.
2. The store-path binary is no longer in the nix binary cache, and the
package is installed via the older way. This should alleviate @lagoja's
concern that there may be a discrepancy between the store-path and the
derivation source in "resolved" field of the package.

## How was it tested?

Did `devbox update` and saw that `go@latest` had new store-paths, and
the "resolved" field was also updated.