[perf] in syncPackagesToProfile, only removeExtraItemsFromProfile is mode is not install (#1541)
## Summary

Doing `devbox global add <package>` was taking 9 seconds for me. Most of
this time
was spent in `removeExtraPackagesFromProfile`. This operation takes
longer
because it calls `devpkg.NormalizedPackageAttributePath` which calls
`nix.Search`
which is slow.

Instead, I advocate for us to not do this step when adding packages.

## How was it tested?

`devbox global add <package>` is now snappy