[telemetry] fix segfault (#1569)
## Summary

`newSentryException` can error
[here](https://github.com/jetpack-io/devbox/blob/main/internal/telemetry/sentry.go#L55)
due to `err` being `nil`. Its only callsite is
[here](https://github.com/jetpack-io/devbox/blob/main/internal/telemetry/telemetry.go#L130),
but the `nix.Version` above is resetting the `err` to `nil` leading to a
segfault.

This was my bad from #1469 

In this PR, I change the variable `err` to `errToLog` so that any of us
adding other code here do not accidentally clobber the `errToLog`
variable.

I think but not sure this is the cause of #1562 (but not sure).

## How was it tested?

I managed to get a repro with the nginx example and running `devbox`
version 0.6, and doing `devbox services up -b`, when the services are
already running leading to an error.

When I run it with the new `devbox` binary, it errors correctly without
a segfault.