devpkg: better flake references and installable parsing (#1581)
In anticipation of adding support for non-default package outputs, I
wanted to clean up some of the flakeref parsing in `devpkg`. This will
make it easier to handle Nix's `flakeref#attrpath^outputs` installable
syntax. This PR is broken up into two commits: one that handles flakeref
parsing and one that handles installable parsing.

The `ParseFlakeRef` function is able to parse the more common flakerefs
that devbox currently supports ("path", "flake", "indirect", etc.). It
doesn't support mercurial, gitlab, and sourcehuthut for now. The
`ParseFlakeInstallable` function parses Nix flake installables, which
are a superset of the flake reference syntax. It allows an attribute
path and/or outputs to be specified using the `flake#attrpath^outputs`
syntax. The tests were validated using the nix expression
`builtins.parseFlakeRef "ref"` and the `nix flake metadata <ref>` CLI.

The goal isn't to perfectly replicate Nix's parsing of flake references
(which itself is sometimes inconsistent), but to make it good enough to
handle flakerefs in devbox.json.