[refactor] Remove devbox interface (#1690)
## Summary

Removes the devbox interface (`devbox.go` in root). I feel like it was
making code harder to follow and not providing any value.

Chatted with @loreto a few weeks ago about doing this. I was about to
add a new function to interface and decided to remove instead.

A possible follow up could be to remove `devopt.Opts` to avoid some
ducplication.

## How was it tested?

builds, CICD.