examples/development/zig: update to 0.11.0 and fix for macOS (#1795)
The zig example is out-of-date and isn't working on macOS. Updating to
the latest version of Zig seems to fix it. I don't think we run these
tests on macOS, so this bug was never caught.

The bug was caused by https://github.com/ziglang/zig/issues/12069 which
has been fixed in 0.11.0.