[cleanup] remove unnecessary code when installing packages in store (#1847)
## Summary

This function used to check for packages in profile, but that's
redundant with
checking in the store and also inaccurate. A package may have been
installed
in the nix store, but not yet in the nix profile: for example, `devbox
add`
but not yet `devbox shell`.

## How was it tested?

`devbox add ripgrep` was already in my store and worked as expected.
`devbox add ripgrep@14.0.2` was not in store and showed output as
expected.