[nix-cache] Add cache copy command (#1891)
## Summary

Note: this may not be the final UX for this command. I'm just getting it
in hidden so we can build the cache feature end-to-end and will refine
details later.

Adds `devbox cache copy <url>` command. It copies the current devbox nix
profile to the url. `<url>` can be anything supported by
https://nixos.org/manual/nix/stable/command-ref/new-cli/nix3-copy

There are two future plans for this command:

* `url` will be optional when logged in. It will be determined by API in
those cases. We may want to move url to be a flag in that case (possibly
the `--to` flag to mimic `nix copy`).
* This command will support specifying the package you wish to copy.
Packages can be nix installable, but also devbox packages (i.e.
package@version). We could even support runx in the future.

I used `copy` to mimic `nix` command name, but I slightly prefer
`upload`. That and the above changes would make this command look like:

`devbox cache upload [--to <url>] [package]`

or as simple as 

`devbox cache upload`

to upload current profile when logged in.

## How was it tested?

`devbox cache copy "s3://mike-test-nix-cache?region=us-west-2"`