[bug-fix] Fix lockfile rewrite (#2007)
## Summary

The source of this bug is that when calling `lockfile.Tidy()` we were
passing in only top level packages and not including plugin packages as
well.

This bug was also affecting dev container construction and some edge
case plugin template logic. This bug only affects users that are using
custom plugins with packages.

I renamed `PackageNames` to `AllPackageNames` for clarity. Also removed
a redundant function in config package that was no longer needed.
 
## How was it tested?

In other project added golang monorepo plugin and logged to ensure Tidy
was not recreating lockfile. Also tested adding go to that project.