[cache] Optimize cache upload (#2083)
## Summary

This changes `devbox cache upload` so that it now longer has to build
packages if they are already in cache.

For each package, it first checks if it is in target cache. If it is, it
skips. Otherwise it builds/copies.

## How was it tested?

<img width="843" alt="image"
src="https://github.com/jetify-com/devbox/assets/544948/225b744a-ff20-446e-a7d0-6436005a33ac">