[nix] Fix StorePathsAreInStore (#2104)
## Summary

Addresses issue introduced in 

https://github.com/jetify-com/devbox/pull/2076
and
https://github.com/jetify-com/devbox/pull/2098

Basically we have 2 different uses cases when looking up store paths.
Sometimes we want all store paths (whether installed or not) and in
other cases we only want paths that are in local store. This change
fixes that and adds some testing.

## How was it tested?

unit tests, manual integration test by installing a package that was not
in local store.