[Remove Nixpkgs] remove feature flag (#2142)
## Summary

We've had this feature launched for quite a while. Initially, I was
reluctant to remove the feature-flag so that we could debug code paths
for users on old versions of nix (which cannot benefit from the binary
cache. IIRC nix < 1.17). However, in the time since I relied on this
exactly once for some perf-related task, and not otherwise.

At this point, I think we should remove it, and even print messages to
users about updating their nix version if needed.

## How was it tested?

It compiles. I could do `devbox shell` in this project.