devpkg: fix empty patch mode (#2278)
Fix a case where an empty patch mode wouldn't default to "auto". This
was causing Devbox to skip patching when it should not have.

Got rid of the sync.OnceValue function for determining if a package
needs to be patched. This was originally done to avoid a call to
nix.System(), but that call doesn't happen anymore.