[init] init config should return error if already exists (#2346)
## Summary

This fixes a NPE in autodetect. Previously when initializing a config
with a directory that already contained a config we would return `(nil,
nil)` but this is a mistake and a bit inconsistent with typical go. This
changes it so that if config already exists we return an error instead.

IMO this also provides better UX displaying a message indicating config
already exists.

This does have a minor backward compatibility issue. If users where used
to calling `devbox init` and having it not fail they would now need to
first test if devbox.json already exists.

## How was it tested?