MySQL Plugin: create and link my.cnf to allow configuring MySQL (#2468)
## Summary

Adds a `$MYSQL_CONF` env var to the MySQL Plugin to allow a custom
`my.cnf` (MySQL config file). MySQL does not provide an environment
variable, command line flag, or anything else to change the location
(it's always computed as `$MYSQL_HOME/my.cnf`), so we simply create a
symlink there pointing to the user's config.

## How was it tested?

- Forked devbox, made changes to the mysql plugin, and ensured
everything worked in a custom project.

---------

Co-authored-by: John Lago <750845+Lagoja@users.noreply.github.com>